impr: chat-list component split + attachment cards
